selectFromDB: Simplify creation of database connection
RPostgres uses libpq, which has full-support for the Pg service and
password files.  Instead of parsing these files ourselves, rely on
standard Pg environment variables to set their paths before making the
connection with just a service name.  This continues to allow swapping
out of the service definition and credentials without changing the R
code.

I wanted to lift the environment variables out into the docker-compose
config, but due to the way the rstudio image works and the context in
which R code runs, that didn't Just Work.  Workarounds could make it
possible, but I'll leave those for the future.